change overall float

+++ App_Daughters/app/src/main/java/momsday/app_daughters/CareworkerInformation/CareworkerInformationContract.java
@@ -2,7 +2,7 @@ package momsday.app_daughters.CareworkerInformation;
 
 public interface CareworkerInformationContract {
     interface View {
-        void setCareworkerInform(String name, String hospitalName, String career, String patientNumber, String introduction, String sincerityScore, String kindnessScore, int overall, String firstReview, String secondReview, String thirdReview);
+        void setCareworkerInform(String name, String hospitalName, String career, String patientNumber, String introduction, String sincerityScore, String kindnessScore, float overall, String firstReview, String secondReview, String thirdReview);
         void showErrorMessage();
     }
