+++ App_Daughters/app/src/main/java/momsday/app_daughters/CareworkerInformation/CareworkerInformationModel.java
@@ -19,7 +19,7 @@ public class CareworkerInformationModel {
     private String[] reviews;
 
     @SerializedName("overall")
-    private int overall;
+    private float overall;
 
     @SerializedName("patientInCharge")
     private int patientNumber;
