+++ App_Daughters/app/src/main/java/momsday/app_daughters/HospitalInformation/HospitalInformationContract.java
@@ -2,7 +2,7 @@ package momsday.app_daughters.HospitalInformation;
 
 public interface HospitalInformationContract {
     interface View {
-        void setHospitalInform(String name, String phoneNumber, String location, String introduction, String facilityScore, String mealScore, String scheduleScore, String costScore, String serviceScore, String firstReview, String secondReview, String thirdReview, int overall);
+        void setHospitalInform(String name, String phoneNumber, String location, String introduction, String facilityScore, String mealScore, String scheduleScore, String costScore, String serviceScore, String firstReview, String secondReview, String thirdReview, float overall);
         void showErrorMessage();
 
     }
